requirements.txt
Flask==2.3.2
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==20.1.0
gevent==22.10.2
